requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
langchain 
transformers 
huggingface-hub
torch
langchain_community